Add new files and the room to display a question
