+++ src/components/collab_elements/codeEditorWindow.jsx
@@ -13,9 +13,9 @@ const CodeEditorWindow = ({ onChange, language, code, theme }) => {
   return (
     <div className="overlay rounded-md overflow-hidden w-full h-full shadow-4xl">
       <Editor
-        height="85vh"
+        height="55vh"
         width={`100%`}
-        language={language || "javascript"}
+        language={language}
         value={value}
         theme={theme}
         defaultValue="// some comment"
